added new Controller and pages view changed BaseController

+++ src/main/java/com/example/spring_boot_project/web/controller/BaseController.java
@@ -9,42 +9,71 @@ import org.springframework.web.bind.annotation.*;
 
 @Controller
 public class BaseController {
-    private static final String REDIRECT = "redirect:/";
+
+    private static final String REDIRECT = "redirect:/users";
     private UserService userService;
 
     public BaseController(UserService userService) {
         this.userService = userService;
     }
 
-    @GetMapping("/")
-    public String printWelcome(ModelMap model) {
+
+    @GetMapping(value = "/users")
+    public String showUsers(ModelMap model) {
         model.addAttribute("users", userService.listUsers());
-        model.addAttribute("add", new User());
-        model.addAttribute("delete", new User());
-        model.addAttribute("update", new User());
         return "users";
     }
 
-    @PostMapping("/")
+    @GetMapping(value = "/add")
+    public String addPP(ModelMap model) {
+        model.addAttribute("add", new User());
+        return "add";
+    }
+
+    @PostMapping("/add")
     public String addUser(@ModelAttribute User user) {
         userService.add(user);
         return REDIRECT;
     }
 
+    @GetMapping(value = "/delete")
+    public String remove(ModelMap model) {
+        model.addAttribute("delete", new User());
+        return "delete";
+    }
+
     @PostMapping("/delete")
     public String removeUser(@ModelAttribute User user) {
+        User us = userService.getUser(user.getId());
+        if (us.getId() == null) {
+            return "redirect:/error";
+        }
         userService.removeUser(user.getId());
         return REDIRECT;
     }
 
-    @PostMapping("/patch")
+    @GetMapping(value = "/update")
+    public String update(ModelMap model) {
+        model.addAttribute("update", new User());
+        return "update";
+    }
+
+    @PostMapping("/update")
     public String updateUser(@ModelAttribute User user) {
         User updateUser = userService.getUser(user.getId());
+        if (updateUser.getId() == null) {
+            return "redirect:/error";
+        }
         updateUser.setName(user.getName());
         updateUser.setSurName(user.getSurName());
         updateUser.setAge(user.getAge());
         userService.add(updateUser);
         return REDIRECT;
+    }
+
+    @GetMapping(value = "/error")
+    public String error() {
+        return "error";
     }
